pyproject.toml
[project]
name = "chat-messaging"
version = "0.1.0"
description = "Messaging service for chat application"
authors = [
    {name = "McEazy2700",email = "codeepoch@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "django-stubs (>=5.1.3,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-filter (>=25.1,<26.0)",
    "markdown (>=3.7,<4.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "drf-yasg[validation] (>=1.21.8,<2.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "drf-nested-routers (>=0.94.1,<0.95.0)",
    "requests (>=2.32.3,<3.0.0)",
    "channels[daphne] (>=4.2.0,<5.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pyright]
reportUnknownVariableType = false
reportUnannotatedClassAttribute = false
reportImplicitOverride = false
reportUnknownMemberType = false
reportMissingTypeStubs = false
reportUnknownArgumentType = false
reportUnknownParameterType = false
reportUnsafeMultipleInheritance = false
reportDeprecated = false
reportAny = false
reportIncompatibleMethodOverride = false
reportMissingSuperCall = false
